fix - refactore the code
