+++ app/views/form-upload.njk
@@ -122,7 +122,7 @@
     <h1 class="govuk-heading-l govuk-!-padding-left-3 govuk-!-margin-bottom-3 govuk-!-margin-top-6">
       2. Evidence
     </h1>
-    <span class="govuk-caption-m">Supported file types are XLS or XLSX. Maximum file size is 20MB.</span>
+    <span class="govuk-caption-m">Supported file types are DOC, DOCX, XLS, XLSX, PDF, JPG, JPEG, PNG, MPG, MP4, WMV, MOV. Maximum file size is 20MB.</span>
     {{
       govukWarningText({
         text: "You must add evidence for each item you are
